styling part for our approach page
