+++ src/components/OurApproach/OurApproach.css
@@ -1,0 +1,11 @@
+.spacing{
+    margin-bottom: 16px;
+    margin-top:32px ;
+}
+
+.spacing1{
+    margin-bottom: 16px;
+}
+.spacing2{
+    margin-bottom: 32px;
+}
